Migrated more player movement code to use floats over shorts.

+++ game/q_shared.h
@@ -759,7 +759,7 @@ typedef struct
 #else
 	short		origin[3];		// 12.3
 #endif	// end Knightmare
-	short		velocity[3];	// 12.3
+	float		velocity[3];	// 12.3
 	byte		pm_flags;		// ducked, jump_held, etc
 	byte		pm_time;		// each unit = 8 ms
 	short		gravity;
